fix: invalid module d ts because computers are horrible
